refactor: move db and controllers in infrastructure folder
