+++ src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './controllers/app.controller';
+import { AppController } from './infrasctructure/controllers/app.controller';
 import { AppService } from './core/services/app.service';
 
 @Module({
