+++ src/infrasctructure/controllers/album.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { AlbumService } from '../core/services/Album.service';
+import { AlbumService } from '../../core/services/Album.service';
 
 @Controller()
 export class AlbumController {
