+++ src/infrasctructure/controllers/app.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get } from '@nestjs/common';
-import { AppService } from '../core/services/app.service';
+import { AppService } from '../../core/services/app.service';
 
 @Controller()
 export class AppController {
